class2_OpenAI_Agents_basics/pyproject.toml
[project]
name = "class2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Farhaj499", email = "farhajshauket@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.10",
    "python-dotenv>=1.1.0",
]

[project.scripts]
class2 = "class2:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
